==> 1_Oscar Best Pictures.ipynb <==
## Oscar Awards Best Picture

#importing libraries
import pandas as pd
import numpy as np
import string
import re

# importing the dataset
oscar_df_1 = pd.read_excel('data/the_oscar_award_kaggle.xlsx')

# Inspecting DataFrame
oscar_df_1.head()
oscar_df_1.shape
oscar_df_1.dtypes
oscar_df_1.isna().sum()

# inspecting the missing values in film\n",
missing_films = oscar_df_1[oscar_df_1.isna().any(axis=1)]
missing_films['category'].value_counts()
#There are no missing values for the best picture category. At this point, there is no need to remove the rows with them.

oscar_df_1['category'].value_counts()

## 2. Creating a dataframe with best picture category

#making a copy of the entire dataset
pictures = oscar_df_1.copy()

#creating categories for filtering
cats = ['OUTSTANDING PICTURE','OUTSTANDING PRODUCTION','OUTSTANDING MOTION PICTURE','BEST MOTION PICTURE','BEST PICTURE']

#selecting only movies that were nominated in the best picture category
pictures = pictures.loc[pictures['category'].isin(cats)]
pictures[pictures.isna().any(axis=1)]
# there are no missing values.

pictures

## 3. Merging Oscar Data With IMDb Movies details

# Importing a csv file with over 80 000 movies and their details from IMDB
IMDB_movies = pd.read_excel('data/IMDB movies extensive dataset kaggle/IMDb_movies.xlsx')

# Inspecting the dataframe

IMDB_movies.shape
# 81273 movies , variables stored in 22 columns

IMDB_movies.head()
IMDB_movies.columns

# dropping unnecessary columns:


IMDB_movies=IMDB_movies.drop(columns = ['original_title','date_published','avg_vote','votes','budget','usa_gross_income',
                                         'worlwide_gross_income', 'metascore','reviews_from_users', 'reviews_from_critics'])

IMDB_movies.head()
IMDB_movies.shape 

#make titles in both dataframes upper-case, to minimize the possible errors caused by different capitalization.
IMDB_movies['title'] = IMDB_movies['title'].str.upper()
pictures['film'] = pictures['film'].str.upper()

#making sure there is a blank space after a comma in names of movies
IMDB_movies['title'] = IMDB_movies['title'].str.replace(',',', ')
pictures['film'] = pictures['film'].str.replace(',',', ')

#Stripping potential blank spaces
IMDB_movies['title'] = IMDB_movies['title'].str.strip()
pictures['film'] = pictures['film'].str.strip()

Most of types look alright.This table will be merged with the pictures table on a left join- so that only the information for the selected oscar films is added

pictures_combined = pd.merge(pictures, IMDB_movies, how='left', left_on=['film', 'year_film'], right_on=['title','year'])

pictures_combined.loc[pictures_combined.film=='LITTLE WOMEN']
# This film from 2019 is wrong in the imdb kaggle dataset. the oscar nominee was Greta Gedwig`s film. it will be manually changed below.

# changing little women info

pictures_combined.at[558,'imdb_title_id'] = 'tt3281548'
pictures_combined.at[558,'director'] = 'Greta Gedwig'
pictures_combined.at[558,'actors'] = "'Saoirse Ronan', 'Emma Watson', 'Florence Pugh', 'Eliza Scanlen','Laura Dern','Meryl Streep'"
pictures_combined.at[558,'director'] = 'Greta Gerwig'
pictures_combined.at[558,'genre'] = "'Drama','Romance'"
pictures_combined.at[558,'duration'] = 135

pictures_combined[pictures_combined['title'].isnull()]

# 6 films were not successfully merged(they were missing in the IMDB dataset). the Main infromation will be added manually:
pictures_combined.at[7,'imdb_title_id'] = 'tt0019257'
pictures_combined.at[7,'genre'] = 'Drama, History, Thriller'
pictures_combined.at[7,'director'] = 'Ernst Lubitsch'
pictures_combined.at[7,'duration'] = 113


pictures_combined.at[14,'imdb_title_id'] = 'tt0021826'
pictures_combined.at[14,'genre'] = 'Drama, Romance'
pictures_combined.at[14,'director'] = 'Frank Lloyd'
pictures_combined.at[14,'duration'] = 102


pictures_combined.at[47,'imdb_title_id'] = 'tt0025986'
pictures_combined.at[47,'genre'] = 'Drama'
pictures_combined.at[47,'director'] = 'Irving Cummings'
pictures_combined.at[47,'duration'] = 80


pictures_combined.at[560,'imdb_title_id'] = 'tt8579674'
pictures_combined.at[560,'genre'] = 'Drama, War'
pictures_combined.at[560,'director'] = 'Sam Mendes'
pictures_combined.at[560,'duration'] = 119

pictures_combined.at[561,'imdb_title_id'] = 'tt7131622'
pictures_combined.at[561,'genre'] = 'Comedy,Drama'
pictures_combined.at[561,'director'] = 'Quentin Tarantino'
pictures_combined.at[561,'duration'] = 161

pictures_combined.at[562,'imdb_title_id'] = 'tt6751668'
pictures_combined.at[562,'genre'] = 'Comedy,Drama, Thriller'
pictures_combined.at[562,'director'] = 'Bong Joon Ho'
pictures_combined.at[562,'duration'] = 132

# Some movies have two directors. Splitting their names in two columns for later checking
pictures_combined[['first_director','second_director']] = pictures_combined.director.str.split(',', expand = True)

pictures_combined.to_csv('data/pictures_combined.csv',index=False)
