==> 2_Scraping_People.ipynb <==
import numpy as np
import string  
import re

from bs4 import BeautifulSoup
import json
import requests

#importing oscar best pictures
pictures_combined = pd.read_csv('data/pictures_combined.csv')

# Creating an empty dataframe for storing the scraped info
best_pictures = pd.DataFrame(columns= ['Name','imdb_name_id','imdb_title_id','Role', 'Sub_Role'])

# 1 SCRAPING CREW OF THE BEST PICTURES

def get_crew(imdb_title_id):

    url = 'https://www.imdb.com/title/'+imdb_title_id+ '/fullcredits'
    header = {'Accept-Language':'en-US;q=0.7, es-MX;q=0.3'}
    req = requests.get(url, headers=header).text
    soup = BeautifulSoup(req)
    credits_table = soup.findAll('div',attrs={'id':'fullcredits_content'})

    for i in credits_table:
        new_row=len(best_pictures)+1
        director_name = i.findAll('td', attrs = {'class':'name'})[0].text.replace('\n', '')
        director_id = i.find('a', href=True).get('href').split('/name/',1)[1][:-1]
       
        best_pictures.loc[new_row,['Name']] = director_name
        best_pictures.loc[new_row,['imdb_name_id']] = director_id
        best_pictures.loc[new_row,['imdb_title_id']] = imdb_title_id
        best_pictures.loc[new_row,['Role']] = 'Director'

        writers_table = i.findAll('table',attrs = {'class':'simpleCreditsTable'})[1].findAll('td',attrs={'class': 'name'})
        producers_table = i.findAll('table', attrs = {'class':'simpleCreditsTable'})[2].findAll('td', attrs = {'class':'name'})                           

        for i in writers_table:

            new_row = len(best_pictures)+1
            writer_name = i.text.replace('\n','')
            writer_id = i.find('a', href=True).get('href').split('/name/',1)[1][:-1]           
            writer_sub_role = i.find_next_siblings()[1].text.replace('\n', '')

            best_pictures.loc[new_row,['Name']] = writer_name
            best_pictures.loc[new_row,['imdb_name_id']] = writer_id
            best_pictures.loc[new_row,['imdb_title_id']] = imdb_title_id
            best_pictures.loc[new_row,['Role']] = 'Writer' 
            best_pictures.loc[new_row,['Sub_Role']] = writer_sub_role             

        for i in producers_table:
            new_row=len(best_pictures)+1
            producer_name = i.text.replace('\n', '')
            producer_id = i.find('a', href=True).get('href').split('/name/',1)[1][:-1]
            producer_sub_role = i.find_next_siblings()[1].text.replace('\n', '')

            best_pictures.loc[new_row,['Name']] = producer_name
            best_pictures.loc[new_row,['imdb_name_id']] = producer_id
            best_pictures.loc[new_row,['imdb_title_id']] = imdb_title_id
            best_pictures.loc[new_row,['Role']] = 'Producer'
            best_pictures.loc[new_row,['Sub_Role']] = producer_sub_role             


for i, imdb_title_id in enumerate(pictures_combined['imdb_title_id']):
    imdb_title_id = pictures_combined.at[i, 'imdb_title_id']    
    url = "https://www.imdb.com/name/"+imdb_title_id+"/"
    req =requests.get(url).text
    soup=BeautifulSoup(req, 'lxml')
    print(i)
    try:
        get_crew(imdb_title_id)
    
    except:
        new_row=len(best_pictures)+1
        best_pictures.loc[new_row,['Name']] = 'EXCEPTION'
        best_pictures.loc[new_row,['imdb_name_id']] = 'EXCEPTION'
        best_pictures.loc[new_row,['imdb_title_id']] = imdb_title_id
        best_pictures.loc[new_row,['Role']] = 'EXCEPTION'
        best_pictures.loc[new_row,['Sub_Role']] = 'EXCEPTION'     
        continue 

best_pictures.loc[best_pictures.imdb_title_id=='tt3281548']

#best_pictures.to_csv('data/best_pictures_WDP.csv', index=False)

#checking which entries were skipped
best_pictures.loc[best_pictures.Name == 'EXCEPTION']

best_pictures1 =best_pictures.drop_duplicates()

best_pictures1.to_csv('data/best_pictures_WDP_withoutduplicates.csv', index=False)

# 2 ADDING DIRECTORS TO THE MAIN DATASET



best_pictures1 = pd.read_csv('data/best_pictures_WDP_withoutduplicates.csv')
pictures_combined = pd.read_csv('data/pictures_combined.csv')

directors = best_pictures1.loc[best_pictures1.Role=='Director']
directors['Name'] = directors['Name'].str.strip()

pictures_combined2=pd.merge(pictures_combined, directors, how='left',on='imdb_title_id')

pictures_combined2[pictures_combined2['first_director']!=pictures_combined2['Name']]
pictures_combined2 = pictures_combined2.rename(columns = {'Name':'Director','imdb_name_id':'director_id'})
pictures_combined2 = pictures_combined2.drop(columns=['Role'])
pictures_combined2 = pictures_combined2.drop(columns=['Sub_Role'])

pictures_combined2.to_csv('data/pictures_combined2.csv',index=False)
