==> 3_Scraping_All_Movies_Directors.ipynb <==
#importing libraries
import pandas as pd
import numpy as np
import string  
import re
# setting the viewing limit for dataframe
pd.set_option("display.max_rows", 101)

from bs4 import BeautifulSoup
import json
import requests

actorProducerDirector = pd.DataFrame(columns= ['Name', 'Movie Name', 'MovieId', 'Year','Role'])

crew = pd.read_csv('data/pictures_combined2.csv')

# getting list of name_ids for each crew member.

crew_list = crew[['director_id','Director']]
crew_df = pd.DataFrame(crew_list).rename(columns={'director_id':'imdb_name_id','Director':'name'})
crew_df = crew_df.drop_duplicates('imdb_name_id')
crew_df = crew_df.reset_index()
crew_df

def fillPerRole(credits, name, typeOfRole):
    role = '';
    if typeOfRole == 'actor':
        role = 'Actor'
    elif typeOfRole == 'director':
        role = 'Director'
    elif typeOfRole == 'producer':
        role = 'Producer'
    
    for credit in credits:
        movieName = credit.find('a', href=True).text
        # HERE I NEED TO ADD SOMETHING TO SKIP TV AND SHORT AND DOCUMENTARIES
        movieId = credit['id'].split('-',1)[1]
        year = credit.find('span', attrs={'class': 'year_column'}).text.replace('\n', '')
        new_row=len(actorProducerDirector)+1
        actorProducerDirector.loc[new_row,['Name']] = name
        actorProducerDirector.loc[new_row,['Movie Name']] = movieName
        actorProducerDirector.loc[new_row,['MovieId']] = movieId
        actorProducerDirector.loc[new_row,['Year']] = year
        actorProducerDirector.loc[new_row,['Role']] = role


def fillActorProducerDirector(nameid):
    header = {'Accept-Language':'en-US;q=0.7, es-MX;q=0.3'}
    url = "https://www.imdb.com/name/"+nameid+"/"
    req = requests.get(url, headers=header).text
    soup = BeautifulSoup(req)
    name = soup.find('div', attrs={'id': 'name-overview-widget'}).find('span', attrs={'class': 'itemprop'}).text

    filmography = soup.findAll('div', attrs={'class': 'filmo-category-section'})
    for film in filmography:
        actorCredits = film.findAll('div', id=lambda value: value and value.startswith('actor'))
        directorCredits = film.findAll('div', id=lambda value: value and value.startswith('director'))
        producerCredits = film.findAll('div', id=lambda value: value and value.startswith('producer'))
        fillPerRole(actorCredits, name, 'actor')
        fillPerRole(directorCredits, name,  'director')
        fillPerRole(producerCredits, name, 'producer')

for i, nameid in enumerate(crew_df['imdb_name_id']):
    name = crew_df.at[i, 'name']    
    url = "https://www.imdb.com/name/"+nameid+"/"
    req =requests.get(url).text
    soup=BeautifulSoup(req, 'lxml')
    print(i)
    fillActorProducerDirector(nameid)


actorProducerDirector

actorProducerDirector.to_csv('temp/temp_actorProducerDirector.csv',index=False)

directors = actorProducerDirector.loc[actorProducerDirector.Role == 'Director']

directors

directors1 = pd.merge(directors,crew_df, how='left', left_on='Name',right_on='name' )

directors1 = directors1.drop(columns=['index','name'])

directors1 = directors1.rename(columns={'Role':'Director'})

directproducers = actorProducerDirector.loc[actorProducerDirector.Role == 'Producer']

directors1.loc[directors1.Name=='Lewis Milestone']

directproducers = directproducers.rename(columns={'Role':'Producer'})
directproducers.Producer = directproducers.Producer.str.replace('Producer','Yes')

directors2 = pd.merge(directors1, directproducers, how='left', on =['Name','Movie Name','MovieId','Year'])

len(directors2)

#dropping duplicated entries

directors2 = directors2.drop_duplicates(subset=['MovieId'])
len(directors2)

directors2 = directors2.fillna('No')

directors2.Year = directors2.Year.str.replace('\xa0', '')

directors2.Year = directors2.Year.str.rstrip('/')
directors2.Year = directors2.Year.str.rstrip('-')

directors2[['Year1','Year2']] = directors2.Year.str.split('-', expand = True)

directors2[['Year3','Year4']] = directors2.Year1.str.split('/', expand = True)

directors2 = directors2.drop(columns=['Year','Year1','Year2','Year4'])

directors2.Year3 = directors2.Year3.str.strip() 

directors2 = directors2.loc[directors2.Year3 != '']

directors2.Year3 = directors2.Year3.astype(int)

directors2 = directors2.rename(columns={'Year3':'Year'})

# spliting directors into ones before 1927 and after, so that we can count how many projects they had.
directors_early = directors2.loc[directors2['Year'] <=1927]
directors_early = directors_early.groupby(['Name','imdb_name_id'])['Count'].sum()
directors_early =directors_early.reset_index()

directors_early['Year'] = 1927

directors_o = directors2.loc[directors2['Year'] >=1928]
directors_o = directors_o.groupby(['Name','imdb_name_id','Year'])['Count'].sum()
directors_o = directors_o.reset_index()

#combining both dataframes
directors_all = directors_early.append(directors_o, ignore_index=True)
directors_all

directors_all_f = directors_all.set_index(["Name",'imdb_name_id','Year']) 
directors_all_f = directors_all_f.groupby(level=[0]).cumsum()
directors_all_f = directors_all_f.reset_index()
directors_all_f = directors_all_f.groupby(['Name','imdb_name_id','Year'])['Count'].max()
directors_all_f = directors_all_f.reset_index()
directors_all_f
directors_all_f = directors_all_f.rename(columns = {'Name':'Dir_Name','Count':'Dir_Cum_Films'})

directors_all_f.columns, crew.columns

pictures_combined3 = pd.merge(crew,directors_all_f, 
                              how='left', 
                              left_on=['director_id','year_film'],
                              right_on=['imdb_name_id','Year'])

pictures_combined3.isna().sum()
pictures_combined3.loc[pictures_combined3.Dir_Cum_Films.isna()]
#8 rows were not merged. They will be added manually

pictures_combined3.at[91,'Dir_Cum_Films'] = 42
pictures_combined3.at[92,'Dir_Cum_Films'] = 68
pictures_combined3.at[98,'Dir_Cum_Films'] = 42
pictures_combined3.at[270,'Dir_Cum_Films'] = 34
pictures_combined3.at[458,'Dir_Cum_Films'] = 18
pictures_combined3.at[472,'Dir_Cum_Films'] = 20
pictures_combined3.at[483,'Dir_Cum_Films'] = 21
pictures_combined3.at[469,'Dir_Cum_Films'] = 14


pictures_combined3.isna().sum()
pictures_combined3 = pictures_combined3.drop(columns=['Dir_Name','imdb_name_id','Year','director'])

len(pictures_combined3)

directproducers.columns
directproducers = directproducers.rename(columns={'MovieId':'imdb_title_id'})
directproducers1 = directproducers[['imdb_title_id','Producer']]
directproducers1 = directproducers1.drop_duplicates(subset=['imdb_title_id'])


#directproducers1.to_csv('temp/directtprodu.csv')

len(directproducers1)

pictures_combined3 = pd.merge(pictures_combined3, directproducers1,
                              how='left',
                             on='imdb_title_id'
                             )

pictures_combined3

pictures_combined3.Producer = pictures_combined3.Producer.fillna('No')

pictures_combined3.to_csv('data/pictures_combined3.csv',index=False)
