==> 4_Expanding_the_Oscars_Dataset_With_Other_Categories.ipynb <==
#importing library
import pandas as pd
import numpy as np

#reading the files
all_oscars = pd.read_excel('data/the_oscar_award_kaggle.xlsx')
pictures_combined3 = pd.read_csv('data/pictures_combined3.csv')

all_oscars['film']=all_oscars['film'].str.upper()
all_oscars['film']=all_oscars['film'].str.strip()

all_oscars['name']=all_oscars['name'].str.strip()

all_oscars.name.loc[all_oscars.name.str.contains('Alejandro')]


# renaming two directors

pictures_combined3['Director'] = pictures_combined3['Director'].str.replace('NoÃ«l Coward','Noël Coward')
pictures_combined3['Director'] = pictures_combined3['Director'].str.replace('Noí«l Coward','Noël Coward')
pictures_combined3['Director'] = pictures_combined3['Director'].str.replace('Alejandro G. IÃ±Ã¡rritu','Alejandro G. Iñárritu')
pictures_combined3['Director'] = pictures_combined3['Director'].str.replace('Alejandro GonzÃ¡lez IÃ±Ã¡rritu','Alejandro G. Iñárritu')
pictures_combined3['Director'] = pictures_combined3['Director'].str.replace('Alejandro G. IÃ±Ã¡rritu','Alejandro G. Iñárritu')


all_oscars['name'] = all_oscars['name'].str.replace('NoÃ«l Coward','Noël Coward')
all_oscars['Director'] = all_oscars['name'].str.replace('Noí«l Coward','Noël Coward')
all_oscars['Director'] = all_oscars['name'].str.replace('Alejandro G. IÃ±Ã¡rritu','Alejandro G. Iñárritu')
all_oscars['name'] = all_oscars['name'].str.replace('Alejandro GonzÃ¡lez IÃ±Ã¡rritu','Alejandro G. Iñárritu')
all_oscars['name'] = all_oscars['name'].str.replace('Alejandro G. IÃ±Ã¡rritu','Alejandro G. Iñárritu')


#counting number of nominations for each film
all_oscars1 = all_oscars[['film','category','ceremony']]
all_oscars1 = all_oscars1.dropna()
all_oscars1['count_of_film_nominations'] = 1
all_oscars1['film']=all_oscars1['film'].str.upper()
all_oscars1['film']=all_oscars1['film'].str.strip()
oscars1 = all_oscars1.groupby(['film','ceremony'])['count_of_film_nominations'].sum()
oscars1 = oscars1.reset_index()
oscars1 = oscars1.sort_values('count_of_film_nominations',ascending = False)


# merging count of nominations with the main dataframe.
pictures_combined3 = pd.merge(pictures_combined3,oscars1,
                             how='left',
                             on=['film','ceremony'])

#checking if everything was merged properly
pictures_combined3.isna().sum()


#checking if everything was merged properly
pictures_combined3.loc[pictures_combined3.count_of_film_nominations.isna()]

#7 movies were not successful. adding them manually:
pictures_combined3.at[92,'count_of_film_nominations'] = 9
pictures_combined3.at[100,'count_of_film_nominations'] = 3
pictures_combined3.at[263,'count_of_film_nominations'] = 4
pictures_combined3.at[267,'count_of_film_nominations'] = 7
pictures_combined3.at[421,'count_of_film_nominations'] = 10
pictures_combined3.at[448,'count_of_film_nominations'] = 6
pictures_combined3.at[545,'count_of_film_nominations'] = 7

pictures_combined3.isna().sum()

pictures_combined3 = pictures_combined3.drop(columns=['second_director','title'])
pictures_combined3

# 1 Adding the Directors Previous Nominations and Wins

# Adding the 'nominated for best director' feature:
all_oscars.category.unique()
best_directing_cat=all_oscars.loc[all_oscars.category.str.contains('DIRECTING')]

best_directing_cat['nominated_for_best_director']= 1

best_directing_cat = best_directing_cat[['ceremony','film','name','nominated_for_best_director','winner']]

best_directing_cat = best_directing_cat.rename(columns={'winner':'best_directing_win'})

pictures_combined3 = pictures_combined3.drop(columns=['name'])

pictures_combined3 = pd.merge(pictures_combined3,best_directing_cat,
                             how='left',
                             on=['film','ceremony'])


pictures_combined3.nominated_for_best_director = pictures_combined3.nominated_for_best_director.fillna(0) 


pictures_combined3[['Dir_Cum_Films','count_of_film_nominations','nominated_for_best_director']]=pictures_combined3[['Dir_Cum_Films','count_of_film_nominations','nominated_for_best_director']].astype(int)

pictures_combined3.columns

### Creating an empty dataframe to add a list of all directors and all years combined

ceremonies=all_oscars.ceremony.tolist()
ceremonies = set(ceremonies)
ceremonies1 = list(ceremonies)


directors = best_directing_cat.name.tolist()
directors=set(directors)
directors =list(directors)


bestpic_directors = pictures_combined3.Director.tolist()
bestpic_directors = set(bestpic_directors)
bestpic_directors = list(bestpic_directors)


all_directors = directors+bestpic_directors
all_directors=set(all_directors)
all_directors = list(all_directors)

from itertools import product

df_directors  = pd.DataFrame(list(product(ceremonies, all_directors)), columns=['ceremony', 'director'])

df_directors

# splitting first and second director
df_directors[['first_dir','second_dir']] = df_directors.director.str.split(',', expand = True)
#getting rid of unecessary columns
df_directors = df_directors.drop(columns=['director','second_dir'])
df_directors

# adding best directing nominations

df_directors1 = pd.merge(df_directors,best_directing_cat,
                        how='left',
                        left_on=['first_dir','ceremony'],
                        right_on=['name','ceremony'])

df_directors1

df_directors1 = df_directors1.rename(columns={'winner':'best_directing_win'})
df_directors1.best_directing_win = df_directors1.best_directing_win.replace({True:1,False:0})

df_directors1 = df_directors1.drop(columns=['name'])

#dropping duplicates and unecessary columns

df_directors1  = df_directors1.drop_duplicates(subset=['ceremony', 'first_dir'], keep='first')

# replacing NaN values with 0
df_directors1=df_directors1.fillna(0)

# Calculating cumulative sums for nominations and wins of directors

noms = df_directors1.copy()

noms1 = noms.set_index(['first_dir','ceremony'])
noms2 = noms1.groupby(level=[0]).cumsum()
noms2 =noms2.reset_index()

noms2 = noms2.rename(columns={'nominated_for_best_director':"best_directing_nomination_sum",
                             'best_directing_win':'best_directing_win_sum'})

noms3 = noms2.sort_values(by=['first_dir','ceremony'])

noms3 = noms3.set_index(['first_dir','ceremony'])

noms4 = noms3.groupby(level=[0]).diff()
noms4 = noms4.reset_index()
noms4 = noms4.fillna(0)

#merging sum of nominations/wins with the difference in order to get prior nominations and wins

directors_final_list = pd.merge(noms2, noms4,
                               how='left',
                               on=['first_dir','ceremony'])

directors_final_list['prior_best_directing_nominations'] = (directors_final_list['best_directing_nomination_sum_x']
                                                            -directors_final_list['best_directing_nomination_sum_y'])
                                                            
                                                            
directors_final_list['prior_best_directing_wins'] = (directors_final_list['best_directing_win_sum_x']
                                                            -directors_final_list['best_directing_win_sum_y'])                                                    

directors_final_list=directors_final_list.rename(columns={'best_directing_nomination_sum_x':'best_directing_nomination_sum',
                                                         'best_directing_win_sum_x':'best_directing_win_sum'})


directors_final_list = directors_final_list.drop(columns=['best_directing_nomination_sum_y','best_directing_win_sum_y'])

#merging directors wiht the best picture dataset

pictures_combined3 = pd.merge(pictures_combined3,directors_final_list,
                             how='left',
                             left_on=['Director','ceremony'],
                             right_on=['first_dir','ceremony'])

pictures_combined3

#adding a category to calculate the number of films the director had nominated for 'best picture' category

total_best_pic = pictures_combined3[['Director','ceremony']]
total_best_pic['moviez'] = 1
total_best_pic = total_best_pic.sort_values(['Director','ceremony'])

total_best_pic = total_best_pic.groupby(['Director','ceremony'])['moviez'].count().reset_index()
total_best_pic1 = total_best_pic.set_index(['Director','ceremony'])
total_best_pic1=total_best_pic1.groupby(level=[0]).cumsum()
total_best_pic1=total_best_pic1.reset_index()
total_best_pic1 = total_best_pic1.rename(columns={'moviez':'running_best_picture_nomination'})

pictures_combined4 = pd.merge(pictures_combined3,total_best_pic1,
                             how='left',
                             on=['Director','ceremony']
                             )

pictures_combined4.columns

pictures_combined4 = pictures_combined4.drop(columns=['name'])
pictures_combined4.best_directing_win = pictures_combined4.best_directing_win.fillna(0)

pictures_combined4

pictures_combined4.to_csv('data/pictures_combined4.csv', index=False)

# 2  Golden Globes

#import previuos version of Oscars dataset. 
pictures_combined4 = pd.read_csv('data/pictures_combined4.csv')

#import dataset from Kaggle
golden_globes = pd.read_csv('data/Golden Globe awards dataset kaggle/golden_globe_awards_K.csv')

# renaming 1917 in oscars dataset
pictures_combined4.film = pictures_combined4.film.str.replace('1917;','1917')

#renaming category'winner' to 'Best_Pic_Winner'
pictures_combined4 = pictures_combined4.rename(columns={'winner':'Best_Pic_Winner'})


#stripping potential blank spaces
golden_globes.film = golden_globes.film.str.strip()
golden_globes.nominee = golden_globes.nominee.str.strip()

#uppercase the film titles
golden_globes.film = golden_globes.film.str.upper()

#renaming columns
golden_globes = golden_globes.rename(columns={'category':'golden_globes_cat','win':'Globes_Winner'})
golden_globes

#renaming once upon a time in hollywood (adding a space after ...)
golden_globes.film = golden_globes.film.str.replace('ONCE UPON A TIME...IN HOLLYWOOD','ONCE UPON A TIME… IN HOLLYWOOD')

## Getting the Nominees for Golden Globes Best Director Category

#creating a dataset with just directors

gg_directors = golden_globes.loc[golden_globes.golden_globes_cat=='Best Director - Motion Picture']
gg_directors = gg_directors.drop(columns=['year_film','ceremony','film'])
gg_directors

#merging gg_directors and oscars

pictures_combined5 = pd.merge(pictures_combined4, gg_directors,
                             how='left',
                             left_on=['year_ceremony','Director'],
                             right_on=['year_award','nominee'])

pictures_combined5 = pictures_combined5.rename(columns={'Globes_Winner':'Globes_Best_Dir_Winner'})
pictures_combined5 = pictures_combined5.drop(columns =['nominee'])

pictures_combined5=pictures_combined5.drop_duplicates()
pictures_combined5

## Adding the nominees for Golden Globes Best Picture Category (incl. drama, comedy and musical)

# Getting the list of films
gg_cats = ['Best Motion Picture - Drama','Best Motion Picture - Musical or Comedy','Picture','Picture - Comedy','Picture - Musical']


gg_pictures = golden_globes.loc[golden_globes.golden_globes_cat.isin(gg_cats)]

gg_pictures = gg_pictures.drop(columns=['film','year_film','ceremony'])
gg_pictures = gg_pictures.rename(columns={'Globes_Winner':'Globes_Best_Pic_Winner'})
gg_pictures.nominee =gg_pictures.nominee.str.upper()

#replacing Hollywood again
gg_pictures.nominee = gg_pictures.nominee.str.replace('ONCE UPON A TIME...IN HOLLYWOOD','ONCE UPON A TIME… IN HOLLYWOOD')

pictures_combined5 = pd.merge(pictures_combined5, gg_pictures,
                             how='left',
                             left_on=['year_ceremony','film'],
                             right_on=['year_award','nominee'])

pictures_combined5 = pictures_combined5.drop(columns = ['year_award_x','golden_globes_cat_x','year_award_y',
                                                       'golden_globes_cat_y','nominee'])

pictures_combined5[['Globes_Best_Dir_Winner','Globes_Best_Pic_Winner']]= pictures_combined5[['Globes_Best_Dir_Winner','Globes_Best_Pic_Winner']].fillna('x')

pictures_combined5.to_csv('data/pictures_combined5.csv', index=False)

# 3 Bafta and Cannes

#importing data

pictures_combined5=pd.read_csv('data/pictures_combined5.csv')

#importing Cannes and Bafta data, obtained through webscraping. Code adjusted from #

bafta = pd.read_csv('data/BAFTA and CANNES datasets/bafta_cleaned.csv')
cannes = pd.read_csv('data/BAFTA and CANNES datasets/cannes_cleaned.csv')

pictures_combined5=pictures_combined5.drop(columns=['year'])

#stripping blank spaces
bafta.film =bafta.film.str.strip()
cannes.film =cannes.film.str.strip()

pictures_combined6 = pd.merge(pictures_combined5, bafta,
                             how='left',
                             left_on=['film','year_film'],
                             right_on =['film','year'] )

len(pictures_combined6)

pictures_combined6= pd.merge(pictures_combined6, cannes,
                             how='left',
                             left_on=['film','year_film'],
                             right_on =['film','year'] )

pictures_combined6 = pictures_combined6.drop(columns=['year_x','year_y']) 

pictures_combined6[['bafta','cannes']]=pictures_combined6[['bafta','cannes']].fillna('x')

# 4 Directors Guild Awards

#importing an excel containing DGA information (obtained through webscraping)
aw = pd.read_excel('data/guild_critic.xlsx')

DGA =  aw.loc[aw.academy == 'DGA Award']
DGA = DGA.rename(columns = {'outcome':'DGA_Winner'})

DGA.title = DGA.title.str.upper()

DGA.name = DGA.name.str.strip()
DGA.title = DGA.title.str.strip()

DGA 

#DGA.name = DGA.name.replace('Alfonso CuarÃ³n','Alfonso Cuarón')

DGA_sels = DGA[['imdb_name_id','DGA_Winner','year']]

pictures_combined6 = pd.merge(pictures_combined6,DGA_sels,
                             how='left',
                             left_on=['year_ceremony', 'director_id'],
                             right_on=['year','imdb_name_id'])

pictures_combined6 = pictures_combined6.drop_duplicates()
pictures_combined6 = pictures_combined6.drop(columns =['imdb_name_id','year'])
pictures_combined6.DGA_Winner=pictures_combined6.DGA_Winner.fillna('x')

# 5 Genre

# Creating columns from genres.
pictures_combined6['Drama'] = np.where(pictures_combined6['genre'].str.contains('Drama'), 1, 0)
pictures_combined6['Adventure'] = np.where(pictures_combined6['genre'].str.contains('Adventure'), 1, 0)
pictures_combined6['War'] = np.where(pictures_combined6['genre'].str.contains('War'), 1, 0)
pictures_combined6['Musical'] = np.where(pictures_combined6['genre'].str.contains('Musical'), 1, 0)
pictures_combined6['Sci-Fi'] = np.where(pictures_combined6['genre'].str.contains('Sci-Fi'), 1, 0)
pictures_combined6['Thriller'] = np.where(pictures_combined6['genre'].str.contains('Thriller'), 1, 0)
pictures_combined6['Romance'] = np.where(pictures_combined6['genre'].str.contains('Romance'), 1, 0)
pictures_combined6['Animation'] = np.where(pictures_combined6['genre'].str.contains('Animation'), 1, 0)
pictures_combined6['Film-Noir'] = np.where(pictures_combined6['genre'].str.contains('Film-Noir'), 1, 0)
pictures_combined6['Horror'] = np.where(pictures_combined6['genre'].str.contains('Horror'), 1, 0)
pictures_combined6['Mystery'] = np.where(pictures_combined6['genre'].str.contains('Mystery'), 1, 0)
pictures_combined6['Music'] = np.where(pictures_combined6['genre'].str.contains('Music'), 1, 0)
pictures_combined6['Action'] = np.where(pictures_combined6['genre'].str.contains('Action'), 1, 0)
pictures_combined6['Crime'] = np.where(pictures_combined6['genre'].str.contains('Crime'), 1, 0)
pictures_combined6['Family'] = np.where(pictures_combined6['genre'].str.contains('Family'), 1, 0)
pictures_combined6['Fantasy'] = np.where(pictures_combined6['genre'].str.contains('Fantasy'), 1, 0)
pictures_combined6['Sport'] = np.where(pictures_combined6['genre'].str.contains('Sport'), 1, 0)
pictures_combined6['Biography'] = np.where(pictures_combined6['genre'].str.contains('Biography'), 1, 0)
pictures_combined6['Comedy'] = np.where(pictures_combined6['genre'].str.contains('Comedy'), 1, 0)
pictures_combined6['History'] = np.where(pictures_combined6['genre'].str.contains('History'), 1, 0)
pictures_combined6['Western'] = np.where(pictures_combined6['genre'].str.contains('Western'), 1, 0)


pictures_combined6.isna().sum()

pictures_combined6.to_csv('data/pictures_combined6.csv', index=False)
